chore(macros): delete legacy js_namespace / js_class proc macros (4.1b)
Every production consumer migrated to holt! / couch! during 4.2a-d.
Strips the legacy proc-macro surface and the supporting types/helpers:

- js_namespace, js_class, js_fn, js_constructor, js_method,
  js_static_method, js_getter, js_setter from
  crates/otter-macros/src/lib.rs (~530 lines including NamespaceArgs
  / ClassArgs / FnArgs / LengthArgs / NameArgs / MethodBinding /
  ConstructorBinding / AccessorBinding + take_*_attr helpers).
- crates/otter-macros/tests/js_namespace.rs
- crates/otter-macros/tests/js_class.rs
- crates/otter-macros/benches/js_surface_macros.rs (compared
  js_namespace / js_class against handwritten — both sides used
  deleted macros, comparison moot).
- criterion dev-dep dropped (no benches left).
- runtime_cx.rs doc references point at holt! / couch! / #[dive].
- Crate-level status block updated to reflect the surface settled
  at holt! / couch! / raft! / #[dive].

Drive-by clippy fix in array_prototype::impl_last_index_of: drop
unnecessary `as usize` cast on a usize value.

Tests: otter-vm 530/530, otter-macros doctests pass, clippy clean.

+++ crates/otter-vm/src/array_prototype.rs
@@ -784,7 +784,7 @@ fn impl_last_index_of(args: &mut IntrinsicArgs<'_>) -> Result<Value, IntrinsicEr
             // `arr.length = 2**31`) does not index out of bounds.
             // Trailing slots beyond `elements.len()` are holes that
             // can never `===` the needle.
-            let mut i = (from as usize).min(elements.len() - 1) as i64;
+            let mut i = from.min(elements.len() - 1) as i64;
             while i >= 0 {
                 if elements[i as usize] == needle {
                     return Some(i as i32);
